refactor(thinking): move ApplyThinkingEffort to protocol/ops
The function was exported from protocol/transform, creating an awkward
cross-package coupling: server/transform_thinking_mode.go had to import
the entire transform package just to call one helper.

Moving ApplyThinkingEffort and its private helpers (disableThinking,
enableThinking, stripOpenAIThinkingExtra, openaiReasoningEffort) to
protocol/ops/request_thinking.go puts them at the right layer — shared
protocol-level pure functions, callable by any consumer without pulling
in TransformContext or chain machinery.

RuleThinkingTransform is now a 4-line thin shell.
ThinkingEffortTransform (server-domain) imports ops directly.
